Fix documentation | Home music | New GameState
We're currently working on a login system for two purposes:
Accessing user data later (a statistic system for when the game is fully built)
Forcing document interaction, so we can play audio elements without error.
-=
Until we fix this issue, to play the home music you first have to generate some sort of interaction with the document. I.E Keyboard click, mouse click, and so forth.