Login System | Fix Errors | Document more |
Okay so to start off, I rewrote how Taylor initiated the Login system. I did this for two reasons:
1.) I found it much easier to hide / remove / show elements at will.
2.) ^ That's how I plan to switch between game states.

The user / pass are : cqb5 / misandryisokay
If you're testing this for your own and want to completely skip the login form, just go into GameState.js and change HOME to true and LOGIN to false. It will remove the login elem, and skip straight to the home scene.

The next step in the Login process (which will be written near the endphase of this project) is to actually create a user stat system / register system. The login function will contact a server to authenticate / store new users..

In the Demo, I'm keeping the Login form there for the very simple reason of it shows off the program in its entirety.